requirements.txt
folium==0.16.0
geopandas==0.14.4
numpy==1.26.4
Pillow==10.3.0
rasterio==1.3.10
rasterstats==0.19.0
streamlit==1.35.0
streamlit_folium==0.20.0
